combine reducer, middleware in recommend.js, check for race condition in middleware
